audio_transcriber/requirements.txt
streamlit==1.38.0
google-generativeai==0.8.1
python-dotenv==1.0.1
pathlib==1.0.1